Initial refactor, still looking to strip main of SQL calls
